Compiler: replace long usages with int32_t

+++ Compiler/script/cc_compiledscript.cpp
@@ -97,7 +97,7 @@ int ccCompiledScript::add_new_function(const char*namm, int *idx) {
     strcpy(functions[numfunctions],namm);
     funccodeoffs[numfunctions] = codesize;
     funcnumparams[numfunctions] = 0;
-    //  code = (long*)realloc(code, codesize + 5000);
+    //  code = (int32_t*)realloc(code, codesize + 5000);
     if (idx)
         *idx = numfunctions;
     numfunctions++;
@@ -183,7 +183,7 @@ int ccCompiledScript::remove_any_import (const char*namm, SymbolDef *oldSym) {
     return 0;
 }
 
-int ccCompiledScript::add_new_export(const char*namm,int etype,long eoffs, int numArgs)
+int ccCompiledScript::add_new_export(const char*namm, int etype, int32_t eoffs, int numArgs)
 {
     if (numexports >= exportsCapacity)
     {
@@ -212,7 +212,7 @@ int ccCompiledScript::add_new_export(const char*namm,int etype,long eoffs, int n
     }
     exports[numexports] = newName;
 
-    export_addr[numexports] = eoffs | ((long)etype << 24L);
+    export_addr[numexports] = eoffs | ((int32_t)etype << 24L);
     numexports++;
     return numexports-1;
 }
